* Conditional 64 bit stat for evfs_fs_posix
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@20362 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/apps/evfs/src/include/evfs_event.h
@@ -66,7 +66,7 @@ typedef struct evfs_stat
 {
    int st_uid;
    int st_gid;
-   int st_size;
+   uint64 st_size;
    int ist_atime;
    int ist_mtime;
    int ist_ctime;
